requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.4.0=h0e60522_1
argon2-cffi=20.1.0=py39h2bbff1b_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39haa95532_0
blas=2.115=mkl
blas-devel=3.9.0=15_win64_mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39hb82d6ee_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.6.15=h5b45459_0
certifi=2022.6.15=py39hcbf5309_0
cffi=1.15.1=py39h0878f49_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=py39haa95532_0
cryptography=37.0.4=py39h7bc7c5c_0
cudatoolkit=11.6.0=hc0ea762_10
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
detecto=1.2.2=pyhd8ed1ab_0
docopt=0.6.2=py_1
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.8=h39d44d4_0
ffmpeg=4.4.2=gpl_ha0349d7_106
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hce3cb01_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.34.4=py39hb82d6ee_0
freeglut=3.2.2=h0e60522_1
freetype=2.10.4=h546665d_1
gettext=0.19.8.1=ha2e2712_1008
glib=2.72.1=h7755175_0
glib-tools=2.72.1=h7755175_0
gst-plugins-base=1.20.3=he07aa86_0
gstreamer=1.20.3=hdff456e_0
icu=70.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.9.1=py39haa95532_0
ipython=8.4.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.33=h77af90b_0
jedi=0.18.1=py39haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
jpeg=9e=h8ffe710_2
jsonschema=4.4.0=py39haa95532_0
jupyter_client=7.2.2=py39haa95532_0
jupyter_core=4.10.0=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.4.4=py39h2e07f2f_0
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=15_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=15_win64_mkl
libclang=14.0.6=default_h77d9078_0
libclang13=14.0.6=default_h77d9078_0
libdeflate=1.12=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libglib=2.72.1=h3be07f2_0
libiconv=1.16=he774522_0
liblapack=3.9.0=15_win64_mkl
liblapacke=3.9.0=15_win64_mkl
libogg=1.3.4=h8ffe710_1
libopencv=4.6.0=py39hbfd931d_3
libpng=1.6.37=h1d00b33_3
libprotobuf=3.20.1=h7755175_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h2ed3b44_1
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.3=h8ffe710_1
libwebp-base=1.2.3=h8ffe710_2
libxcb=1.13=hcd874cb_1004
libxml2=2.9.14=hf5bbc77_3
libzlib=1.2.12=h8ffe710_2
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib-base=3.5.2=py39h581301d_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.4.4=py39haa95532_0
nbformat=5.3.0=py39haa95532_0
nest-asyncio=1.5.5=py39haa95532_0
notebook=6.4.12=py39haa95532_0
numpy=1.23.1=py39h1a62c8c_0
opencv=4.6.0=py39hcbf5309_3
openh264=2.2.0=h0e60522_2
openjpeg=2.4.0=hb211442_1
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py39h2e25243_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h0e60522_0
pdf2image=1.16.0=pyhd8ed1ab_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39ha53f419_0
pip=22.1.2=py39haa95532_0
pipreqs=0.4.11=pyhd8ed1ab_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.6.0=py39h832f523_3
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39hcbf5309_5
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.12.0=py3.9_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywin32=302=py39h2bbff1b_2
pywinpty=2.0.2=py39h5da7b33_0
pyzmq=23.2.0=py39hd77b12b_0
qt-main=5.15.4=h467ea89_2
requests=2.28.1=pyhd8ed1ab_0
scikit-learn=1.1.1=pypi_0
scipy=1.9.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.5=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
svt-av1=1.1.0=h0e60522_1
tbb=2021.5.0=h2d74725_1
terminado=0.13.1=py39haa95532_0
testpath=0.6.0=py39haa95532_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h8ffe710_0
torchvision=0.13.0=py39_cu116
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=hda174b7_0
ucrt=10.0.20348.0=h57928b3_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.29.30037=h902a5da_6
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39hcbf5309_4
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
x264=1!161.3030=h8ffe710_1
x265=3.5=h2d74725_3
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
yarg=0.1.9=py_1
zeromq=4.3.4=hd77b12b_0
zlib=1.2.12=h8ffe710_2
zstd=1.5.2=h6255e5f_2
